Add Nack support

+++ input/internal/reader/reader.go
@@ -1,28 +1,31 @@
 package reader
 
 import (
+	"container/list"
 	"context"
 	"dhickie/redpanda-connect-sqs-fifo/input/internal/aws"
 	"dhickie/redpanda-connect-sqs-fifo/input/internal/models"
 	"dhickie/redpanda-connect-sqs-fifo/input/internal/tracking"
-	"slices"
 	"time"
 )
 
 type SqsFifoReader struct {
-	client     *aws.SqsClient           // A client for the SQS API
-	tracker    *tracking.MessageTracker // Tracks all in-flight messages
-	pendingAck []*string                // Messages that have been acknowledged by the runtime but not yet deleted
-	conf       *models.InputConfig      // The configuration for the input
+	client      *aws.SqsClient           // A client for the SQS API
+	tracker     *tracking.MessageTracker // Tracks all in-flight messages
+	pendingAck  *list.List               // Messages that have been acknowledged by the runtime but not yet deleted
+	pendingNack *list.List               // Messages that have had a negative acknowledgement by the runtime but not yet processed
+	conf        *models.InputConfig      // The configuration for the input
 }
 
+// TODO add max capacity to all slices where possible
 func NewSqsFifoReader(conf *models.InputConfig) *SqsFifoReader {
 	client := &aws.SqsClient{}
 	return &SqsFifoReader{
-		client:     client,
-		tracker:    tracking.NewMessageTracker(conf, client),
-		pendingAck: make([]*string, 0), // TODO add max capacity based on max in flight
-		conf:       conf,
+		client:      client,
+		tracker:     tracking.NewMessageTracker(conf, client),
+		pendingAck:  list.New(),
+		pendingNack: list.New(),
+		conf:        conf,
 	}
 }
 
@@ -48,7 +51,12 @@ func (r *SqsFifoReader) Next() *models.SqsMessage {
 
 // Ack acknowledges a message and adds it to the list of pending acknowledgements
 func (r *SqsFifoReader) Ack(id *string) {
-	r.pendingAck = append(r.pendingAck, id)
+	r.pendingAck.PushBack(id)
+}
+
+// Nack acknowledges a message has failed processing and adds it to the list of pending failed acknowledgements
+func (r *SqsFifoReader) Nack(id *string) {
+	r.pendingNack.PushBack(id)
 }
 
 // Reads messages from the queue if there's room in the buffer
@@ -74,29 +82,49 @@ func (r *SqsFifoReader) readLoop() {
 func (r *SqsFifoReader) ackLoop() {
 	t := time.NewTicker(time.Second)
 	for range t.C {
-		// Get the actual messages
-		msgs := make([]*models.SqsMessage, 0, len(r.pendingAck))
-		for _, id := range r.pendingAck {
-			msg, err := r.tracker.Peek(id)
+		batch := make([]*models.SqsMessage, 0, 10)
+		for e := r.pendingAck.Front(); e != nil; e = e.Next() {
+			msg, err := r.tracker.Peek(e.Value.(*string))
 			if err != nil {
 				// TODO log an error here
 				continue
 			}
-			msgs = append(msgs, msg)
+
+			batch = append(batch, msg)
+			if len(batch) == 10 || e.Next() == nil {
+				err := r.client.DeleteMessages(context.TODO(), batch)
+				if err != nil {
+					// TODO log an error here
+					// TODO deal with failed batch members
+				}
+
+				for _, bMsg := range batch {
+					r.tracker.Ack(bMsg.Msg.MessageId)
+				}
+
+				clear(batch)
+			}
 		}
+	}
+}
 
-		// Batch them into 10 max messages
-		batches := slices.Chunk(msgs, 10)
-		for batch := range batches {
-			err := r.client.DeleteMessages(context.TODO(), batch)
+// TODO add thread safety around pending collections
+func (r *SqsFifoReader) nackLoop() {
+	t := time.NewTicker(time.Second)
+	for range t.C {
+		ids := make([]*string, 0, r.pendingNack.Len())
+		for e := r.pendingNack.Front(); e != nil; e = e.Next() {
+			id := e.Value.(*string)
+			_, err := r.tracker.Peek(id)
 			if err != nil {
 				// TODO log an error here
 				continue
 			}
 
-			for _, msg := range batch {
-				r.tracker.Ack(msg.Msg.MessageId)
-			}
+			// We don't batch here for Nacks because the tracker will need to batch again anyway when resetting
+			// message visibility for additional messages in the group
+			ids = append(ids, id)
+			r.tracker.Nack(ids...)
 		}
 	}
 }
